CPR-199 move message sending helper for reuse
